requirements.txt
python-fasthtml
uvicorn
numpy
scipy
networkx
scikit-fuzzy
pydantic
python-multipart
fastlite